pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ledger-cfo"
version = "0.1.0"
description = "A CFO agent for managing ledger data, potentially interacting with Google Ads."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", # Assuming MIT, update if needed
    "Operating System :: OS Independent",
]
dependencies = [
    # Core dependencies identified
    "click", # For CLI
    "pytest", # For tests
    "flask", # For web server in __main__
    "sqlalchemy", # For database interactions (CRUD)
    "psycopg2-binary", # Assuming PostgreSQL, adjust if different DB used
    "python-quickbooks", # For QBO API
    "cachetools", # For caching
    "google-api-python-client", # For Gmail API
    "google-auth-httplib2", # For Google Auth
    "google-auth-oauthlib", # For Google Auth
    "openai", # For LLM interaction
    "anthropic", # For Claude fallback
    "python-dotenv", # For loading .env files
    "tenacity", # For retries
    "google-cloud-secret-manager", # For accessing secrets
    "cloud-sql-python-connector[psycopg2]", # For Cloud SQL connection
    # Add google-ads if confirmed needed
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "flake8",
    # Add other dev dependencies
]

[project.urls]
"Homepage" = "https://github.com/Sku3397/ledger-cfo" # Update if needed
"Bug Tracker" = "https://github.com/Sku3397/ledger-cfo/issues" # Update if needed

[project.scripts]
ledger-cfo = "ledger_cfo.cli:main"

[tool.setuptools.packages.find]
where = ["src"]  # Look for packages in src/ 